Simple ChatBot/requirements.txt
streamlit
langchain
langchain-google-genai
google-generativeai
python-dotenv
pypdf
